Listen on radiation.messages topic, event service cleanup and move each event view to separate file
