--- lib/helpers/links_helper.ex
@@ -1,5 +1,0 @@
-defmodule LinksHelper do
-  def artwork_link(artwork_id) do
-    "https://www.artsy.net/artwork/#{artwork_id}"
-  end
-end
