+++ lib/services/summary_service.ex
@@ -5,7 +5,7 @@ defmodule Aprb.Service.SummaryService do
   def update_summary(topic, event) do
     current_date = Calendar.Date.today! "America/New_York"
     cond do
-      Enum.member?(~w(subscriptions users inquiries purchases conversations), topic.name) ->
+      Enum.member?(~w(subscriptions users inquiries purchases conversations radiation.messages), topic.name) ->
         handle_summary(topic, event["verb"], current_date)
       topic.name == "bidding" ->
         handle_summary(topic, event["type"], current_date)
