+++ lib/views/user_slack_view.ex
@@ -1,0 +1,10 @@
+defmodule Aprb.Views.UserSlackView do
+  import ViewHelper
+
+  def render(event) do
+    %{
+      text: ":heart: #{cleanup_name(event["subject"]["display"])} #{event["verb"]} #{artist_link(event["properties"]["artist"]["id"])}",
+      unfurl_links: true
+    }
+  end
+end
